refactor: split error code into separate file
