+++ packages/aql/src/compiler/compiler.ts
@@ -1,4 +1,5 @@
 import { QueryState } from "../query";
+import { CompileError, getCompileError, saveStack } from "./errors"
 
 let _uid = 0;
 function resetUid() {
@@ -8,14 +9,6 @@ function resetUid() {
 function uid(tableName) {
   _uid++;
   return tableName + _uid;
-}
-
-class CompileError extends Error {
-  constructor(message: string = "") {
-    super(message);
-
-    Object.setPrototypeOf(this, CompileError.prototype);
-}
 }
 
 function nativeDateToInt(date) {
@@ -367,83 +360,6 @@ function validateArgLength(arr: unknown[], min: number, max: number | null = nul
   if (max != null && arr.length > max) {
     throw new CompileError('Too many arguments');
   }
-}
-
-//// Nice errors
-
-function saveStack(type, func) {
-  return (state, ...args) => {
-    if (state == null || state.compileStack == null) {
-      throw new CompileError(
-        'This function cannot track error data. ' +
-          'It needs to accept the compiler state as the first argument.'
-      );
-    }
-
-    state.compileStack.push({ type, args });
-    let ret = func(state, ...args);
-    state.compileStack.pop();
-    return ret;
-  };
-}
-
-function prettyValue(value) {
-  if (typeof value === 'string') {
-    return value;
-  } else if (value === undefined) {
-    return 'undefined';
-  }
-
-  let str = JSON.stringify(value);
-  if (str.length > 70) {
-    let expanded = JSON.stringify(value, null, 2);
-    return expanded.split('\n').join('\n  ');
-  }
-  return str;
-}
-
-function getCompileError(error, stack) {
-  if (stack.length === 0) {
-    return error;
-  }
-
-  let stackStr = stack
-    .slice(1)
-    .reverse()
-    .map(entry => {
-      switch (entry.type) {
-        case 'expr':
-        case 'function':
-          return prettyValue(entry.args[0]);
-        case 'op': {
-          let [fieldRef, opData] = entry.args;
-          return prettyValue({ [fieldRef]: opData });
-        }
-        case 'value':
-          return prettyValue(entry.value);
-        default:
-          return '';
-      }
-    })
-    .map(str => '\n  ' + str)
-    .join('');
-
-  const rootMethod = stack[0].type;
-  const methodArgs = stack[0].args[0];
-  stackStr += `\n  ${rootMethod}(${prettyValue(
-    methodArgs.length === 1 ? methodArgs[0] : methodArgs
-  )})`;
-
-  // In production, hide internal stack traces
-  if (process.env.NODE_ENV === 'production') {
-    const err = new CompileError();
-    err.message = `${error.message}\n\nExpression stack:` + stackStr;
-    err.stack = null;
-    return err;
-  }
-
-  error.message = `${error.message}\n\nExpression stack:` + stackStr;
-  return error;
 }
 
 //// Compiler
